Fixed triangle filter calculation
The issue was caused by precision loss in LinqExtensions.AverageTime, ending in weird results depending on the given times

Also added some changes so that the unit tests can use the FilterSample class which is the expected API to be used. This is due to considering removing a redundant calculation in TriangeFilter.