added current sensing noise treshold to config
Since different boards (AC / DC) and loads give different current sensing noise levels, an explicit treshold has been added to the config.

The default is the highest we have seen, but when switching heaters with smaller currents this must be changed.

Many recurrent on entries in the log are a sign this value is too high for the target loads, while many recurrent off messages are likely due to noise.