removing the workaround given latest info
Reason: the reason the issue still triggered without the delay between processes, is because the conflict also involved a process started by a third party API being used on the affected systems. So the delay is what in reality was avoiding the conflict with that process and the lock was not doing anything for that.

Better fix for our case: the program that runs in the other systems will be adjusted so that the offending third party code runs at a time where we can be certain no other child processes are being ran by us.