fixed crash when ignoring boards
Reason: when reconnect attempts to a board has been exceeded the code ignores the board when BoardSettings.StopWhenLosingSensor is set to false. The problem was that the code ignoring the board was removing it from a list shared by the code reporting the connection states sent in the vector. Which ended in an error like the following: System.ArgumentException: wrong vector length (input, expected): 34 <> 35