Attempt at fixing log spamming due to time out writing to switchboards / part 2
Inner logic was still writing a failure on every timeout, instead of letting the new lower frequency logging to run. It meant a message on every cycle on cases where writing to the board would permanently fail.

Added the time between the timeout and the first succesful write to the later message.