avoids tight look if the TryReadLineDelegate keeps throwing an exception
Reason: not flagging the buffer as examined causes ReadAsync to return inmediately so the recurring failures would have nothing to throttle their speed. By flagging it as examined it will only try as often as we get data or when hitting a timeout. In general TryReadLineDelegate should avoid throwing as it will still be a problem, but this avoids making the situation worse.