resuming/stopping switchboard actuations now only show in the logs
reason: these are only valuable for postmortem analysis, as read loops already report messages for disconnects, stale communications detected and other types of errors. A few temporary situations are not reported to the operator other than through the plots state, but this is intended and can still be seen in logs for postmortem analysis.