moved odd workaround in HandleCommand to its correct location
this removes

context: when handling user input we explicitely write the entered characters in the screen, but we were missing writing the newline.

known issue: non user commands will no longer be forced into a separate line vs. user input, but anyway it was not a good a experience. Correct usability way would be for the non user output/command to show *above* the text the user was typing, so that the text the user typed so far always shows together at the bottom.