fixed json deserialization of SystemChangeNotificationData
reason: apparently having the setter of SystemChangeNotificationData.NodeName as internal set did not work properly for apps packages as single app release apps. For example, local attempts at simple unit test reproduction did not show the problem, but on the target system debugging showed conclusively the json had the data and failed to be included in the parsed result.