removed early shutdown of heaters when reaching target temperature
reason: more standard p control behavior, as we execute the full actuation decided in the control cycle and then adjust what to do in the next cycle. If/when we change p control to actually control the power level, then we would not need the time check in MustTurnOff (as we would sustain the power level until the next actuation cycle).