simplified building subsystems actuations
specifically added BaseSensorBox.RegisterWriteActions which is a simpler API that sends commands to the box for a given IOconfOutput.

See the updated SwitchBoardController for an updated example of this in practice. Do note however that that example does not show the decision logic, as that is found in the related HeatingController.