improved command feedback
Breaking Change: the oven command with multiple areas has been removed. This is how it now works:
- `oven singletarget` is still supported
- `ovenarea all singletarget` does the same thing
- ovenarea area target

It is now checked that the full decision command matches + an optional target.

Note this mainly validates the command format is correct, but the decision logic can have other additional reasons to reject a command, including only the range of values accepted for the target. Check the plots to see if the system started executing the command.