fixed  #207
The problem: we have one instance of BaseSensorBox x type of sensor. Each instance with custom writes rejects the command if its not one of its boards.

This was fixed by backporting 4.x changes to the DefaultCommandRunner so it no longer aborts when one of the command handlers rejects the command. Instead it runs all the command handlers and reports the command as accepted if at least one handler accepted the command.