fixed: potentially slow reaction on nodes with multiple devices
It is expected the issue could cause slow downs of 0.1 to 1 second depending on the amount of devices used on a node and whether alarms are also in use.

root cause: each vector was only delivered to one of the subsystem, so they would take turns handling different vectors.

Note this is a regression in 4.x, so 3.x is unnaffected.