added support for GenericOutput
This allows to tie any vector field to an output by specifyng a simple template in a config line.

This makes it possible to have subsystems that are done purely on plugins + config lines.

The generic aspect here means not specific to a board type. However, without host customizations the support is limited to boards using the standard copenhagen atomics communications format (usb serial port with 115200 baud rate that returns csv lines at 10Hz + receives commands as a line of text)