added OvenProportionalControlUpdates (#259)
* added OvenProportionalControlUpdates

This allows live updates to the proportional control arguments of the heating control. Note that the updated values are lost when restarting the system.

The configuration line requires specifying the maximum value for the different proportional control arguments. The meaning and format of the arguments is the same as in the Oven configuration lines.

The feature is only supported at the oven area level, so all the heaters in the area get the same values. When the arguments have not been changed, the system uses the configuration values specified in the Oven configuration lines for each heater.

* IO.conf lines:
    * OvenProportionalControlUpdates;[MaxProportionalGain];[MaxControlPeriod];[MaxOutputPercentage] e.g. OvenProportionalControlUpdates;10;00:01:00;100
* user commands:
    * set proportional gain for oven area 2: ovenarea 2 pgain 0.5
    * set control period for oven area 1: ovenarea 1 controlperiodseconds 100
    * set max output for oven area 3: ovenarea 3 maxoutput 70
* plugin fields: ovenarea{number}_pgain, ovenarea{number}_controlperiodseconds, ovenarea{number}_maxoutput. The fields are state fields that retain their value once set. The change is only reflected in the next control period.

* fixed: stackoverflow with negative control period seconds

Both the pgain and max output now treat negative numbers as 0 and thus keep the heater off.
A negative control period is treated as a 0.1 seconds control period, so that the state of the heater is re-evaluated on every decision cycle.