subsystems can now report global inputs (#260)
On distributed deployments, the inputs are normally associated with a specific node. This change allows to have inputs that are not specific to a given node.

For example, this can enable a leader based host to report timing of operations done by the currently elected leader. The alternative without the new feature is to add every leader field to all nodes and only to report a value for the leader, which not only is confusing but wastes a lot of resources by needing so many extra/unused fields.