timeout changes for vports (#262)
It now expects 10 seconds between reads (we might turn it into a setting eventually)

It now does not attempt reconnects until an hour passes. This is a workaround to socat exiting when the serial port connection is closed when the reconnect logic runs.

If a vport does receive data for 1 hour, LoopControl must be restarted for the vport to work again.