added initial sensor redundancy support (#268)
* added initial sensor redundancy support

The redundancy support can be used to treat a group of sensors as if it was only one sensor for related control functions. To do this, the system uses the median of all the specified sensors. Sensors are skipped from the median calculation if connection to the sensor's box or node is lost.

IOconf lines:
* RedundantSensors;name;sensor1;sensor2;...;sensorn
* RedudantValidRange;name;min;max i.e. if a sensor gives values outside the range it is considered defective or disconnected and skipped from the median calculation
* RedundantInvalidDefault;name;invaliddefaultvalue i.e. specifies a value to use if we can't get data from any of the sensors; choose a value that would cause the related control function to go into its default state